Inventory two tabs
