+++ src/components/Inventory_Components/index.js
@@ -1,0 +1,1 @@
+export { default as Raw } from './Raw';
